Pull BGAPI command packet builder into a separate class.
